refactor: move `ecma` test to `uglifyOptions` test
